Split `item_missing` into multiple crate pairs
