+++ test_crates/enum_missing/new/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "enum_missing"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
