+++ test_crates/enum_missing/new/src/main.rs
@@ -1,0 +1,3 @@
+pub mod my_pub_mod {
+    pub enum PubUseRemovedEnum {}
+}
