+++ test_crates/enum_missing/old/src/main.rs
@@ -1,0 +1,7 @@
+pub enum WillBeRemovedEnum {}
+
+pub mod my_pub_mod {
+    pub enum PubUseRemovedEnum {}
+}
+
+pub use my_pub_mod::PubUseRemovedEnum;
