+++ test_crates/function_missing/new/src/main.rs
@@ -1,0 +1,3 @@
+pub mod my_pub_mod {
+    pub fn pub_use_removed_fn() {}
+}
