+++ test_crates/function_missing/old/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "function_missing"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
