+++ test_crates/struct_missing/new/src/main.rs
@@ -1,0 +1,3 @@
+pub mod my_pub_mod {
+    pub struct PubUseRemovedStruct;
+}
