+++ test_crates/struct_missing/old/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "struct_missing"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
