+++ test_crates/struct_missing/old/new/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "template"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
