+++ test_crates/struct_missing/old/src/main.rs
@@ -1,0 +1,7 @@
+pub struct WillBeRemovedStruct;
+
+pub mod my_pub_mod {
+    pub struct PubUseRemovedStruct;
+}
+
+pub use my_pub_mod::PubUseRemovedStruct;
