+++ test_outputs/function_missing.output.ron
@@ -2,21 +2,21 @@
     {
         "name": String("will_be_removed_fn"),
         "path": List([
-            String("item_missing"),
+            String("function_missing"),
             String("will_be_removed_fn"),
         ]),
         "visibility_limit": String("public"),
         "span_filename": String("src/main.rs"),
-        "span_begin_line": Uint64(7),
+        "span_begin_line": Uint64(1),
     },
     {
         "name": String("pub_use_removed_fn"),
         "path": List([
-            String("item_missing"),
+            String("function_missing"),
             String("pub_use_removed_fn"),
         ]),
         "visibility_limit": String("public"),
         "span_filename": String("src/main.rs"),
-        "span_begin_line": Uint64(29),
+        "span_begin_line": Uint64(4),
     }
 ]
