+++ test_outputs/inherent_method_missing.output.ron
@@ -3,24 +3,24 @@
         "name": String("Foo"),
         "method_name": String("will_be_removed_associated_fn"),
         "path": List([
-            String("item_missing"),
+            String("inherent_method_missing"),
             String("Foo"),
         ]),
         "visibility_limit": String("public"),
         "method_visibility": String("public"),
         "span_filename": String("src/main.rs"),
-        "span_begin_line": Uint64(15),
+        "span_begin_line": Uint64(7),
     },
     {
         "name": String("Foo"),
         "method_name": String("will_be_removed_method"),
         "path": List([
-            String("item_missing"),
+            String("inherent_method_missing"),
             String("Foo"),
         ]),
         "visibility_limit": String("public"),
         "method_visibility": String("public"),
         "span_filename": String("src/main.rs"),
-        "span_begin_line": Uint64(17),
+        "span_begin_line": Uint64(9),
     }
 ]
