+++ test_outputs/struct_missing.output.ron
@@ -3,23 +3,23 @@
         "name": String("WillBeRemovedStruct"),
         "struct_type": String("unit"),
         "path": List([
-            String("item_missing"),
+            String("struct_missing"),
             String("WillBeRemovedStruct"),
         ]),
         "visibility_limit": String("public"),
         "span_filename": String("src/main.rs"),
-        "span_begin_line": Uint64(3),
+        "span_begin_line": Uint64(1),
     },
     {
         "name": String("PubUseRemovedStruct"),
         "struct_type": String("unit"),
         "path": List([
-            String("item_missing"),
+            String("struct_missing"),
             String("PubUseRemovedStruct"),
         ]),
         "visibility_limit": String("public"),
         "span_filename": String("src/main.rs"),
-        "span_begin_line": Uint64(25),
+        "span_begin_line": Uint64(4),
     },
     {
         "name": String("UnitStructToPlain"),
